requirements.txt
django==1.8.2
django-foundation-icons==3.1
django-bourbon==3.1.8
django-autoconfig==0.2.8
django-jquery==1.9.1
django-compressor==1.5
django-foundation-statics==5.4.7
django-appconf==1.0.1
django-casper==0.0.3
djangorestframework==3.1.3
unittest2==1.0.1
pyyaml==3.11
six==1.9.0
docutils==0.12
Pillow==2.8.2
